+++ packages/adminui-vue/CHANGELOG.md
@@ -33,7 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
       }
       AdminUiVue.mountHelpers.mount(mountParams)
     }
-  ```     
+  ```
+ - added `UForm` component. Wrap for UFormRow which can set label width for child `UFormRow`s
 ## [1.4.0]
 ### Changed
  - UbContext -> UContextMenu